refactor(_shared): move HealthTierExt out of wifi_ui to its own domain
HealthTierExt.resolveLabel() lived in wifi_ui.dart (a WiFi-signal display
helper), which forced every health-score consumer to import the WiFi domain
just to render a tier label — an SRP/layering violation flagged in PR #979
review. Move the extension next to HealthTier in network_health_helpers.dart
and drop the now-unneeded wifi_ui imports from the 5 affected files.

Also document the three look-alike tier enums (NodeSignalLevel, SignalTier,
HealthTier) with their data source, purpose, and a "do NOT merge" note plus
cross-references, since their overlapping value names invite accidental merging.
